+++ src/io/output/FileWriter.h
@@ -2,7 +2,7 @@
 #define FILEWRITER_H
 
 #include <string>
-#include <particleSim/ParticleContainer.h>
+#include <simulator/ParticleContainer.h>
 #include <memory>
 
 namespace output{
